[NRRPLT-5675] Remove limitations for distributed experiments.
This patch removes explicit limitations for the transfer function and
brain editors and enables reset functionality for distributed
experiments. These functions are fully supported by the new distributed
nest implementation, but will fail and cause the simulation to terminate
for MUSIC based experiments.

There is no way for the frontend to distinguish between a pure nest or
MUSIC-nest experiment, so this is the only viable solution while still
supporting both types of distribution.

Change-Id: I5dabdef2dc991d3ddef72a2602c774c5515b83b3